Add build script for generating documentation
We only discovered that the Sphinx autodoc extension imports the modules
listed in `import` statements (and similar) when processing Python source
files. (See: http://sphinx-doc.org/latest/ext/autodoc.html)

This means that we will need to have the modules present for generating
documentation, including those from 3rd party libraries. We were not
getting failures previously due to these libraries available
system-wide.

This patch features a build script for generating the documentation, and
is a one command solution doing so. Yep, we know humans are lazy.

Signed-off-by: Pang Yan Han <yanhan@viki.com>