Use `viki_fabric_config.yml` for configuration for viki-fabric-helpers
This patch fixes the location of the YAML file for the
viki-fabric-helpers package to `viki_fabric_config.yml`. This makes it
easier for future patches which may add new keys to the file for
configuring other parts of the package.

Reading of the configuration file happens at the `viki.fabric` package
level, when any module underneath the `viki.fabric` package is imported.

Signed-off-by: Pang Yan Han <yanhan@viki.com>